fix(rangeiter): use float division for even step distribution in LinearIterator (#7895)
## Summary
- Fixed integer division truncation in `NewLinearIterator` step size
calculation that caused the last step interval to be disproportionately
larger than all other intervals.
- Changed `float64(totalRange / (stepCount - 1))` to
`float64(totalRange) / float64(stepCount - 1)` so the fractional
remainder is distributed evenly across all steps instead of accumulating
in the last one.
- Added a new test case that specifically covers non-divisible ranges to
prevent regression.

## Details

The step size calculation performed integer division before converting
to float64:
```go
// Before (integer division truncates remainder)
stepSize := float64(totalRange / (stepCount - 1))

// After (proper float division preserves remainder)
stepSize := float64(totalRange) / float64(stepCount - 1)
```

For example, with `min=0, max=100, stepCount=7`:
- **Before:** steps `[0, 16, 32, 48, 64, 80, 100]` — gaps of
16,16,16,16,16,**20** (last gap 25% larger)
- **After:** steps `[0, 16, 33, 50, 66, 83, 100]` — gaps of
16,17,17,16,17,17 (evenly distributed)

This affects `DynamicTaskBatchSizer` in the replication queue processor,
where uneven steps can cause unexpected jumps when scaling batch sizes.

Fixes #7894

## Test plan
- [x] Updated existing test expectations for the large negative range
case to match correct float division behavior
- [x] Added new test case `"even distribution with non-divisible range"`
that verifies even step distribution for range [0, 100] with 7 steps
- [x] All existing `rangeiter` tests pass: `go test
./common/rangeiter/... -v -count=1`
- [x] `go vet ./common/rangeiter/...` clean

+++ common/rangeiter/linear_iterator.go
@@ -73,7 +73,7 @@ func NewLinearIterator[T Integer](min, max T, stepCount int) *LinearIterator[T] 
 	}
 
 	// Calculate step size as a float to ensure even distribution
-	stepSize := float64(totalRange / (stepCount - 1))
+	stepSize := float64(totalRange) / float64(stepCount-1)
 
 	// Generate steps
 	steps := make([]T, stepCount)
